refactor: Remove unused contact support module
