+++ backend/api/quivr_api/models/settings.py
@@ -4,6 +4,7 @@ from uuid import UUID
 from langchain.embeddings.base import Embeddings
 from langchain_community.embeddings.ollama import OllamaEmbeddings
 from langchain_community.vectorstores.supabase import SupabaseVectorStore
+
 from langchain_openai import OpenAIEmbeddings
 from posthog import Posthog
 from pydantic_settings import BaseSettings, SettingsConfigDict
@@ -18,6 +19,13 @@ logger = get_logger(__name__)
 class BrainRateLimiting(BaseSettings):
     model_config = SettingsConfigDict(validate_default=False)
     max_brain_per_user: int = 5
+
+
+
+class ContactsSettings(BaseSettings):
+    model_config = SettingsConfigDict(validate_default=False)
+    resend_contact_sales_from: str = "null"
+    resend_contact_sales_to: str = "null"
 
 
 # The `PostHogSettings` class is used to initialize and interact with the PostHog analytics service.
