+++ backend/api/quivr_api/modules/tools/email_sender.py
@@ -11,8 +11,7 @@ from langchain_community.document_loaders import PlaywrightURLLoader
 from langchain_core.tools import BaseTool
 from pydantic import BaseModel
 from quivr_api.logger import get_logger
-from quivr_api.models.settings import BrainSettings
-from quivr_api.modules.contact_support.controller.settings import ContactsSettings
+from quivr_api.models.settings import BrainSettings, ContactsSettings
 from quivr_api.packages.emails.send_email import send_email
 
 logger = get_logger(__name__)
